+++ internal/database/queries/creatures.sql
@@ -33,7 +33,7 @@ SELECT
     alignment,
     challenge_rating,
     embedding_model,
-    (1 - (embedding <=> $1)) as similarity
+    (1 - (embedding <=> $1))::float8 as similarity
 FROM scry_quest.bestiary
 WHERE embedding IS NOT NULL
 ORDER BY embedding <=> $1
